-- more fixes
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@15 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/src/protocols/rtp/connectivity/outboundconnectivity.cpp
@@ -340,7 +340,7 @@ bool OutboundConnectivity::FeedVideoDataUDP(msghdr &message) {
 		GETCLOCKS(rtpDouble);
 		//		ASSERT("_pOutStream->GetCapabilities()->audioCodecInfo.aac.sampleRate: %d",
 		//				_pOutStream->GetCapabilities()->audioCodecInfo.aac.sampleRate);
-		put_htonl(buff + 16, (rtpDouble / CLOCKS_PER_SEC)
+		put_htonl(buff + 16, (uint32_t)(rtpDouble / CLOCKS_PER_SEC)
 				* _pOutStream->GetCapabilities()->audioCodecInfo.aac.sampleRate); //RTP ts
 		put_htonl(buff + 20, _videoPacketsCount); //sender's packet count
 		put_htonl(buff + 24, _videoBytesCount); //sender's octet count
@@ -398,7 +398,7 @@ bool OutboundConnectivity::FeedAudioDataUDP(msghdr &message) {
 		put_htonll(buff + 8, ntp); //NTP
 		double rtpDouble;
 		GETCLOCKS(rtpDouble);
-		put_htonl(buff + 16, (rtpDouble / CLOCKS_PER_SEC)*1000 * 90); //RTP ts
+		put_htonl(buff + 16, (uint32_t)(rtpDouble / CLOCKS_PER_SEC)*1000 * 90); //RTP ts
 		put_htonl(buff + 20, _audioPacketsCount); //sender's packet count
 		put_htonl(buff + 24, _audioBytesCount); //sender's octet count
